requirements.txt
fastapi 
uvicorn 
python-multipart
openai 
langchain
langchain-community
faiss-cpu 
pypdf
tiktoken
python-dotenv